Changes: * Replace the main app files with a unique file 'app.py', this makes the code easier to read. * Create layout with a function invocation with the required args in 'app.py', instead of importing the args from 'app.py' to the layout files, this avoids circular import errors derived from the 1st change. * Move code related to API key retrieval to api.py.
