+++ api.py
@@ -1,11 +1,15 @@
 import datetime as dt
 import os
+from pathlib import Path
 
 import pandas as pd
 import requests
+from dotenv import load_dotenv
 from forex_python.converter import CurrencyRates
 
 
+env_file = Path(__file__).resolve().parent / '.env'
+load_dotenv(env_file)
 POLYGON_API_KEY = os.environ.get('POLYGON_API_KEY')
