+++ layout/tab_sections/fng.py
@@ -1,150 +1,153 @@
-import datetime as dt
 import dash_daq as daq
 import dash_bootstrap_components as dbc
 from dash import html, dcc, dash_table
 
-from callbacks import df_fng_sampled
+from utils import resample_df_fng
 
 
-fng_gauge_table = (
-    html.Section(
-        children=[
-            html.Div(
-                daq.Gauge(
-                    id='fear_greed_index',
-                    color={
-                        "gradient": True,
-                        "ranges": {
-                            "red": [0, 33],
-                            "yellow":[33, 66],
-                            "green":[66, 100]
-                        }
-                    },
-                    value=int(df_fng_sampled['Value'].loc[0]),
-                    showCurrentValue=True,
-                    label='Fear and Greed Index ',
-                    max=100,
-                    min=0,
+def render_fng_table(df_fng):
+    df_fng_sampled = resample_df_fng(df_fng)
+    fng_gauge_table = (
+        html.Section(
+            children=[
+                html.Div(
+                    daq.Gauge(
+                        id='fear_greed_index',
+                        color={
+                            "gradient": True,
+                            "ranges": {
+                                "red": [0, 33],
+                                "yellow":[33, 66],
+                                "green":[66, 100]
+                            }
+                        },
+                        value=int(df_fng_sampled['Value'].loc[0]),
+                        showCurrentValue=True,
+                        label='Fear and Greed Index ',
+                        max=100,
+                        min=0,
+                    ),
+                    className='fng-part-data'
                 ),
-                className='fng-part-data'
-            ),
-            html.Div(
-                dash_table.DataTable(
-                    data=df_fng_sampled.to_dict('records'),
-                    columns=[
-                        {"name": i, "id": i}
-                        for i in df_fng_sampled.columns
-                    ],
-                    style_header={
-                        'backgroundColor': 'rgb(30, 30, 30)',
-                        'color': 'white',
-                        'text-align': 'center',
-                        'display': 'none'
-                    },
-                    style_data={
-                        'backgroundColor': 'rgb(50, 50, 50)',
-                        'color': 'white',
-                        'text-align': 'center'
-                    },
-                    style_cell_conditional=[
-                        {
-                            'if': {'column_id': 'Value'},
-                            'width': '70px'
+                html.Div(
+                    dash_table.DataTable(
+                        data=df_fng_sampled.to_dict('records'),
+                        columns=[
+                            {"name": i, "id": i}
+                            for i in df_fng_sampled.columns
+                        ],
+                        style_header={
+                            'backgroundColor': 'rgb(30, 30, 30)',
+                            'color': 'white',
+                            'text-align': 'center',
+                            'display': 'none'
                         },
-                        {
-                            'if': {'column_id': 'Label'},
-                            'width': '150px'
+                        style_data={
+                            'backgroundColor': 'rgb(50, 50, 50)',
+                            'color': 'white',
+                            'text-align': 'center'
                         },
-                        {
-                            'if': {'column_id': 'Time'},
-                            'width': '100px'
-                        }
-                    ],
-                    style_data_conditional=[
-                        {
-                            'if': {
-                                'filter_query': '{Value} > 0 && {Value} <= 25',
-                                'column_id': 'Value'
+                        style_cell_conditional=[
+                            {
+                                'if': {'column_id': 'Value'},
+                                'width': '70px'
                             },
-                            'backgroundColor': 'tomato',
-                            'color': 'white'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} > 25 && {Value} < 50',
-                                'column_id': 'Value'
+                            {
+                                'if': {'column_id': 'Label'},
+                                'width': '150px'
                             },
-                            'backgroundColor': 'rgba(206, 140, 104, 1)',
-                            'color': 'white'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} = 50',
-                                'column_id': 'Value'
+                            {
+                                'if': {'column_id': 'Time'},
+                                'width': '100px'
+                            }
+                        ],
+                        style_data_conditional=[
+                            {
+                                'if': {
+                                    'filter_query': '{Value} > 0 && {Value} <= 25',
+                                    'column_id': 'Value'
+                                },
+                                'backgroundColor': 'tomato',
+                                'color': 'white'
                             },
-                            'backgroundColor': 'rgb(220, 220, 7)',
-                            'color': 'white'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} > 50 && {Value} <= 75',
-                                'column_id': 'Value'
+                            {
+                                'if': {
+                                    'filter_query': '{Value} > 25 && {Value} < 50',
+                                    'column_id': 'Value'
+                                },
+                                'backgroundColor': 'rgba(206, 140, 104, 1)',
+                                'color': 'white'
                             },
-                            'backgroundColor': 'rgb(41, 183, 41)',
-                            'color': 'white'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} > 75 && {Value} <= 100',
-                                'column_id': 'Value'
+                            {
+                                'if': {
+                                    'filter_query': '{Value} = 50',
+                                    'column_id': 'Value'
+                                },
+                                'backgroundColor': 'rgb(220, 220, 7)',
+                                'color': 'white'
                             },
-                            'backgroundColor': 'rgb(8, 130, 8)',
-                            'color': 'white'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} > 0 && {Value} <= 25',
-                                'column_id': 'Label'
+                            {
+                                'if': {
+                                    'filter_query': '{Value} > 50 && {Value} <= 75',
+                                    'column_id': 'Value'
+                                },
+                                'backgroundColor': 'rgb(41, 183, 41)',
+                                'color': 'white'
                             },
-                            'color': 'tomato'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} > 25 && {Value} < 50',
-                                'column_id': 'Label'
+                            {
+                                'if': {
+                                    'filter_query': '{Value} > 75 && {Value} <= 100',
+                                    'column_id': 'Value'
+                                },
+                                'backgroundColor': 'rgb(8, 130, 8)',
+                                'color': 'white'
                             },
-                            'color': 'rgba(206, 140, 104, 1)'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} = 50',
-                                'column_id': 'Label'
+                            {
+                                'if': {
+                                    'filter_query': '{Value} > 0 && {Value} <= 25',
+                                    'column_id': 'Label'
+                                },
+                                'color': 'tomato'
                             },
-                            'color': 'rgb(220, 220, 7)'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} > 50 && {Value} <= 75',
-                                'column_id': 'Label'
+                            {
+                                'if': {
+                                    'filter_query': '{Value} > 25 && {Value} < 50',
+                                    'column_id': 'Label'
+                                },
+                                'color': 'rgba(206, 140, 104, 1)'
                             },
-                            'color': 'rgb(41, 183, 41)'
-                        },
-                        {
-                            'if': {
-                                'filter_query': '{Value} > 75 && {Value} <= 100',
-                                'column_id': 'Label'
+                            {
+                                'if': {
+                                    'filter_query': '{Value} = 50',
+                                    'column_id': 'Label'
+                                },
+                                'color': 'rgb(220, 220, 7)'
                             },
-                            'color': 'rgb(8, 130, 8)'
-                        },
-                    ]
-
+                            {
+                                'if': {
+                                    'filter_query': '{Value} > 50 && {Value} <= 75',
+                                    'column_id': 'Label'
+                                },
+                                'color': 'rgb(41, 183, 41)'
+                            },
+                            {
+                                'if': {
+                                    'filter_query': '{Value} > 75 && {Value} <= 100',
+                                    'column_id': 'Label'
+                                },
+                                'color': 'rgb(8, 130, 8)'
+                            },
+                        ]
+                    ),
+                    className='fng-part-data'
                 ),
-                className='fng-part-data'
-            ),
-        ],
-        className='main-fng-box'
+            ],
+            className='main-fng-box'
+        )
     )
-)
+    return fng_gauge_table
+
+
 fng_selector_graph = (
     html.Section(
         children=[
@@ -171,6 +174,8 @@ fng_selector_graph = (
         className='graph-container'
     )
 )
+
+
 fng_info_button = (
     html.Div(
         children=[
